fix cities lists feature create city from settings and mark city as a favorite

+++ app/src/androidTest/java/com/example/meteoapp/data/CityDaoTest.kt
@@ -56,7 +56,7 @@ class CityDaoTest {
     @Test
     fun testGetCityByName() = runBlocking {
         val city = cityDao.getCityByName(testCity.name).toLiveData(1)
-        assertThat(getValue(city)[0]!!.id.toLong(), equalTo(testCityId))
+        assertThat(getValue(city)[0]!!.id, equalTo(testCityId))
     }
 
     @Test
